Commit everything in its current state
Want to start tracking this stuff, so we commit everything as it is.
This way, some of these commented out things and experiments can be
removed from the latest version of stuff, and things can hopefully get a
bit simpler.

Also hoping to combine all the dotfiles across machines. But we'll see
how that goes.