Start using `flake-parts`
Apparently, `flake-utils` is accidentally popular, while `flake-parts`
is intentionally popular. This click-bait article actually explains it
fairly well: https://ayats.org/blog/no-flake-utils.

The long and short of that article is that `flake-utils` is a handful of
useful helpers that have to be used correctly. It doesn't go much into
how `flake-parts` is better aside from stating that it will type check
the `mkFlake` helper. There's more explanation on https://flake.parts.

We're not really using `flake-parts` all that much just yet. This is
mostly converting over the flake that already existed.