Use `nixfmt-rfc-style` for `.nix` files
Apparently, this is the soon-to-be defacto formatter. Might as well get
on board with it now.

This falls into the same trap most formatters fall into: caring about
line length. Line length in a formatter causes noisy diffs, plan and
simple. You reword a line, and if that line happens to be part of a
larger expression (which happens very often) you get a huge diff for
adding a couple of characters.

Maybe someone will make a new formatter that doesn't care about line
length, but instead is a Vonderhaar-style formatter. But that's a battle
I can't win even in a language I'm more involved in.

Or maybe the `git` world finally realizes that semantic diffs are more
valuable than syntactic diffs. So we don't have to worry about it
anyway.

Until something happens, we use these formatters that care about line
length :|.