Setup `treefmt-nix`
This seems like it'll be easier to manage all the different formatters
if we use `treefmt`, as that's its purpose.

There's a `flake-part` for it, so we pull it in and use it.

`treefmt` behaves…a little odd. It purports to take a `--stdin` command,
but still wants a filename for that command. We cannot give it some
arbitrary filename, because it will complain if it cannot find it.
Because of this, we can't use it in a tool that wants things to happen
with STDIN/STDOUT only.

It also wants to traverse the entire file system if you use it as
`treefmt`. That's definitely not what we want. We don't want a mistype
of `treefmt` (instead of `treefmt .`) to format files in some completely
different directory.

Finally, it's noisy as heck. Every file that doesn't have a formater
ends up prining a warning. There doesn't seem to be a way to turn that
off aside from explicitly excluding things. It would be nice for it to
respect what's in the `.gitignore` file, so we don't have to manage the
same information in both places.

For these reasons, we don't add it to the `devShell`. It wouldn't really
be useful. It will stay relegated as something the tools use. They can
set up the environment to work properly. And we'll use `nix fmt` as the
interface for formatting.