Start extracting `git` stuff to a module
There's a non-trivial amount of setup for `git` stuff. Some of the more
convoluted bits are signing commits/tags. The only thing you really need
is the signing key and email, and you should be able to generate
basically everything else.

We extract a module that does basically that. If this commit is signed,
then it worked.