Disable mandb cache
This is a long time coming. Apparently, the reason Home Manager builds
were taking multiple minutes was that it was rebuilding the `mandb`
cache and that's slow:
https://github.com/NixOS/nixpkgs/issues/100288#issuecomment-709328701.

We disable it and it speeds up rebuilds tremendously. Like from 2.5
minutes down to 20 seconds.

Also, it seems like `apropos` and `man` still auto-complete. So it's
unclear what this was doing. Maybe a restart will show the difference.
Either way, I personally don't do too much `apropos`/`man`ing. So it
seems fine for now.