Setup `vim-surround` for IdeaVim
We want to use `vim-surround` in JetBrains IDEs. the IdeaVim plugin
doesn't suppor that out of the box. You have to depend on that plugin
yourself.

So what we're doing here is setting up `vim-plug` to pull down `vim`
plugins, pulling down `vim-surround`, and enabling it in the
`~/.ideavimrc` file.

With that all setup, we should be able to start using `vim-surround`.

It'd be nice to manage all this through `nix`, but it wasn't working
right for some reason. Instead, we dish out to `vim-plug` to pull things
down.