Add some stuff for a private repository
This should all be in the `flake.nix` file of the private repository,
but it's out of date and too much effort to update at the moment. So
these tools are installed globally. Once that private repository is
up to date, we can push this stuff there, where it belongs.