Setup `mac-app-util`
This is amazing! We can have actual macOS Applications installed from
`nix` visible to Spotlight! Been wanting this for years.

There aren't many things that are currently in this configuration that
produce macOS Applications. In part, because they weren't usable before.
But now that this is setup, it's time to start tracking more stuff in
`nix`!