Setup Godot for macOS
There's not a package for Godot on macOS in `nixpkgs`. It's not clear if
that's an inertia thing, since the original change only supported Linux:
https://github.com/NixOS/nixpkgs/pull/21689. Or if there's something
actually preventing Godot from being built on macOS.

Regardless, we don't build things from source in this repo, so we pull
in the package directly from the releases on macOS. For Linux, we use
what's in `nixpkgs`.

In any case, this seems to work. And with `mac-app-util`, it behaves as
though it's installed like normal.

We also put the module and package in the outputs, so this can be used
in other places (like individual project repositories).