Setup Godot with .NET support for macOS
This is quite the dance we're doing, all because the Godot developers
are being stubborn about how they want .NET to be used:
https://github.com/godotengine/godot-proposals/issues/1941. This is
another issue that's been open for multiple years that has a solution,
but isn't being implemented.

It needs the .NET SDK to either be in one of a few hard coded paths, or
it needs the `DOTNET_ROOT` and `PATH` environment variables to be set.
Honestly, those are two fine ways to resolve the .NET SDK. But the
problem is that those are the only two ways to resolve the .NET SDK.

This makes it nearly impossible to have the macOS app work like any
other app on macOS unless you install the .NET SDK globally. When you
attempt to open it, it will throw an error modal up because it cannot
find the .NET SDK.

The workaround is to setup a couple Launchd agents to set the
`DOTNET_ROOT` and `PATH` environment variables at login time. This makes
it so GUI applications launched the normal way on macOS (Spotlight,
double clicking in Finder, etc.) all get this environment variable set.
Hopefully this doesn't break something else that's looking for this
`DOTNET_ROOT` variable, nor does it pollute the `PATH` too much.

It's incredibly wild that this is the only way around this issue. But
this seems to be par for the course.

In any case, this seems to work. So we can start using Godot with .NET
support! We also put this package in the outputs, so we can use it as
well in other places.