Simplify `godot-mono` setup
Instead of doing this dance with the Launchd agents, we make a
trampoline much like `mac-app-util` does, and set the environment
variable in there.

This means that the actual Application Bundle is now usable as-is. So we
can use it directly without needing to wrap anything else.

Unfortunately, `osacompile` looks to be a macOS only thing. There
doesn't seem to be an equivalent package in `nixpkgs`. So we can only
build this package on macOS machines. In practice, it shouldn't mean
much; since this is realistically only something that's running on one
machine anyway. But in theory, it means we can't really share this
wider (like to upstream it in `nixpkgs`). Maybe a better solution will
come along in the future.