Use an even simpler method for `godot-mono`
We take apage from `nixpkgs` itself, with things like
`vscode-with-extensions`:
https://github.com/NixOS/nixpkgs/blob/e1b2ab8d1ed56608d2ede877fab5bbe2a15e75fe/pkgs/applications/editors/vscode/with-extensions.nix#L84-L94.

The long and short is that we symlink everything except the actual
binary. Then wrap it in a script that sets the `DOTNET_ROOT` and `PATH`
environment variables.

This is basically what was happening before, except less work. It also
doesn't depend on the build platform being Darwin all so that it can run
`osacompile`. Which means we should be able to build this on any
platform!