Copy Application Bundles but symlink the contents
As it turns out, the only thing that actually needs to be a real file is
the Application Bundle directory itself. The rest of the stuff can be
symlinked, and it all seems to work out. This was originally suggested
in https://github.com/hraban/mac-app-util/issues/27.

The actual thing that Home Manager does is fairly trivial. It creates a
bunch of files that are from the output of `buildEnv`. Because they're
created with the `home.file` module, Home Manager will clean them up
when they're no longer around.

What we're doing here is first inlining the behavior from Home Manager
as an intermediate output. These are symlinks to the actual Application
Bundles in the Nix store. We take each of these symlinks, create an
actual Application Bundle directory of the same name, and symlink all of
the contents in the symlinked Application Bundle into the newly created
Application Bundle. That's the whole trick.

With this new module doing everything we wanted from `mac-app-util`, we
remove that as a dependency. `mac-app-util` does more stuff, but we
don't use that behavior at the moment. We might add it back as a
dependency in the future, if that changes.

More to the point, this feels like it should be upstreamed into Home
Manager proper. It's a relatively small change, but it seems to make
everything Just Work™. We'll see what the outcome of the discussion is
in https://github.com/hraban/mac-app-util/issues/27.