Pull in applications from `homebrew-casks`
Since we can have macOS Application Bundles work like they're installed
normally, we can start tracking Application Bundles we cannot get
through `nixpkgs` from `homebrew-casks` instead!

We use `brew-nix` to pull in Casks. There's another slightly older
project called `nixcasks` which works simlarly. But it doesn't support
PKG files. Not really clear how many Casks use a PKG file, but I'd
rather start with something that supports them.

`brew-nix` doesn't seem to support `x86_64-darwin` out of the box, since
all of the `url`s point to the `aarch64-darwin` versions. So, we have to
do a little munging of the data to make it work on `x86_64-darwin`.

In any case, this opens up a whole world of options for what we can
track in Nix. Not every Cask from `homebrew-casks` will work, since
`brew-nix` doesn't translate every possible behavior of a Cask to Nix.
But we can get a chunk of stuff tracked.

We start with:
- `arduino-ide` for building embedded things.
    - Mostly ESP32 stuff, these days.
- `buckets` for personal finance management.
    - Haven't really used this in years, though.
- `calibre` for book and document management.
- `chromium` for a different web browser when the others don't support
    some API (like WebSerial or some such).
- `discord` for using Discord outside of a browser.
- `duckduckgo` for experimenting with another browser.
- `elgato-stream-deck` for controlling the physical Elgato Stream Deck.
- `firefox` for a privacy-conscious browser.
- `freecad` for 3D modeling.
- `handbrake` for ripping physical DVDs and BDs to secondary storage.
- `jetbrains-toolbox` for those open source projects I can use JetBrains
    on.
    - This is mostly a springboard to using Fleet, but Fleet support is
        still lacking these days.
- `kicad` for 3D modeling.
- `krita` for 2D design.
    - It's an open source competitor to the Adobe suite.
- `makemkv` for transcoding those ripped DVDs and BDs to more manageable
    storage.
    - We don't have a ton of DVDs/BDs, but going from multiple gigs per
        disc to under a gig helps with storage space.
- `mqtt-explorer` for diagnosing MQTT behavior.
- `obsidian` for knowledge management.

There are still a handful of Applications that are in the
`/Applications` directory, but don't work through Nix just yet. We'll
get to them in the future.