Allow overridding .NET SDK for `godot-mono`
This is a confusing situation. The .NET SDK isn't a build-time
dependency of `godot-mono`, it's a runtime dependency. So `godot-mono`
deciding which attribute to use doesn't really make sense aside from
wanting to have _something_ chosen so that it can work out of the box.
Ultimately, the decision of which .NET SDK to use depends on the
specific use of `godot-mono` at a given point in time.

But another way, if a program is being built with the .NET SDK at
version 8.0 one day, and then it needs to use 10.0 the next day; nothing
about `godot-mono` really needs to change except it needs to update its
`DOTNET_ROOT` and `PATH` variables to point to the new .NET SDK. But
neither of those are inherent to building `godot-mono`, they just make
it so people don't have to figure out these hard to find out values and
set them themselves each time they want to use `godot-mono`.

We buck the trend a bit by adding an attribute as an argument that
doesn't actually exist, and let that be the `override-dotnet-sdk`
attribute. The reasoning behind this is explained more in the comment.
But the gist is that `dotnet-sdk` isn't recommended, so we don't want to
use that as an attribute. `dotnet-sdk` is also a real package, and that
would probably make things more confusing if we did use that as an
attribute.

In any case, this seems to work. This even works if we combine multiple
.NET SDKs (though it's unclear if Godot proper likes that).