Provide all packages in Flake output
While we cannot fix the modules getting out of sync immediately, we can
fix the packages getting out of sync. There's a helper in
`nixpkgs`–`lib.filesystem.packagesFromDirectoryRecursive` that does
basically what we want, but we ned to migrate from using `default.nix`
to using `package.nix` for all of the packages.

This lets us have all the current packages, and any future packages
automatically be part of the Flake's output.

Now to sort out the modules…