Put all modules in outputs automatically
We follow a similar pattern to how we manage putting all packages in the
outputs automatically for modules. It's a little different, since
modules don't need to be called like a package. But it's effectively
the same deal.

This should make it so things don't get out of sync when we add a new
module in the future.