Allow specifying the Application Bundle directory
The naming here is because I don't want to come back at this in the
future and assume it's the behavior of Home Manager. I want it to be
explicitly not something they would do. But on other systems, that's
probably not wanted.

We make it possible to set the directory where the Application Bundles
are copied.