Restructure git SSH signing
Move this to a structure similar to what it might look like upstream.

Also be explicit that this is SSH signing, and not GPG signing. There
was a thought that we could support either. But honestly, GPG is such an
archaic system that unless something forces it, going back to it seems
unlikely.