Reorganize Fish plugins
We move the `nix-env.fish` module to be under `programs.fish` as though
it might exists upstream.

The difference here is that `programs.fish.plugins` already exists, and
it is a submodule. So we cannot easily extend it. Instead, we make a
different attribute: `programs.fish.package-plugins` and put
`nix-env.fish` there.

We also adjust the attribute to be `"nix-env.fish"` instead of
`nix-env_fish`. Mostly for consistency with the actual plugin name.