Make a helper for `lib.options.mkPackageOption`
We've written this logic enough times to warrant pulling it out to its
own helper. This should make things a bit easier to deal with in the
future.

We also update the `lib.options.mkPackageOption` uses that reference
packages within their `pkgs` argument. Being consistent will probably be
easier in the long run than having to know when to use one or the other.