Don't force `programs.git.enable` to `true`
You should still be able to control whether `programs.git.enable` is
`true` from the `home.nix` file. But before, `programs.git.ssh-signing`
was forcing `programs.git.enable` to always be `true` whenever
`programs.git.ssh-signing.enable` was `true`. That's not right, since it
means that overriding that becomes harder.

We remove this hardcoded value, so we don't have to go through a bunch
of steps to disable `programs.git.enable`.