Make a `programs.git.structural` module
`difftastic` is really nice to use when trying to understand better what
the diff actually is. But, it's not quite ready to be used full time.
And that's for two reasons (neither of which are its own fault):
1. Not all `git` commands can use a separate diff tool.
    - E.g. `git add --patch`.
    - Or if they can use a separate diff tool, it's unclear how to use
        one.
2. Most every other computer and person doesn't understand the format.
    - Computers for sure don't know how to display the diff well if you
        were to copy/paste it somewhere.
    - For people, it's more a lack of familiarity and preference than
        anything else.

Regardless, we don't want to use `difftastic` for everything right now.
Instead, we add a new module that tosses in a few aliases so that
`difftastic` use can be opted-in when appropriate.