Add busted ole TurboTax 2024
Hate giving Intuit any sort of support, but this one is out of my hands.
I don't even remember if the `dontStrip` is necessary, or if I was too
impatient to wait for it. And honestly, I don't care. Everything about
this is awful. But I'd rather track it, than not.