Revert "Setup `sapling`"
This reverts commit 209236b2374377857b9be4e07c0b64d2948bed58.

`sapling` is tightly coupled to `gh`. That by itself is fine, but the
issue is that too many things are hardcoded to the literal characters
`gh`. Because of this, the 1Password setup won't work for `sapling`. And
without that, ISL doesn't work. It's unclear if anything else works,
because I didn't put in the time to figure it out.

Between the authentication issues, and only supporting GPG signing,
`sapling` isn't ready for me to use. Maybe sometime in the future when
its more fleshed-out.