Generalize commit signing module
We're about to setup `jujutsu`, and it uses most of the same information
for commit signing. We shouldn't have to re-invent this logic for a
different program.

More to the point, both `git` and `jujutsu` need the same
`~/.ssh/allowed-signers` file. Attempting to create that file in two
different places wouldn't work well for `nix`.

We'll add support for `jujutsu` shortly.