Update `restack` hashes
It's completely opaque _why_ these needed to be updated. But the
previous hashes weren't correct. It doesn't make a whole lot of sense,
as they seem to match up with _something_:
```Console
$ nix-prefetch-url --unpack https://github.com/abhinav/restack/releases/download/v0.8.0/restack-darwin-amd64.tar.gz | xargs nix hash to-sri --type sha256
path is '/nix/store/v3xfmjjiirk99ihyzx9rz3pavhw2v20q-restack-darwin-amd64.tar.gz'
sha256-9kkPnTGPnye5PLTMm1lBMJ783lFqZ56B4kPyEbyoz64=
```

Regardless, this stuff needs to be updated in order to bump the flake
inputs, so we do it.

We also take the moment to move from `sha256` to `hash`, as the `sha256`
attribute seems to be deprecated.