Move `homeConfigurations` into `perSystem`
As it turns out, trying to use `homeConfigurations` for the same
username with a different `system` is overly hard:
https://github.com/nix-community/home-manager/issues/3075.

Because the configuration needs the `pkgs` to be resolved, it means that
each username decides which system it's going to be used on. If you
wanted to use the same username on a different system, that won't work
unless you jump through a bunch of hoops. You can either:
- Slap on a bunch of extra stuff onto `home-manager switch`,
    `nh home switch`, or whatever, so that it will resolve the correct
    username. This is super error prone, and you now have to know too
    many implementation details just to use `home-manager`.
- Try to use the `<username>@<hostname>` pattern for the username. If
    the hostname is the same on both systems, this doesn't actually do
    anything extra over just using the username of `<username>`. It also
    doesn't work for machines where the hostname is not known a priori
    (e.g. if you subscribe to cattle not pets:
    https://cloudscaling.com/blog/cloud-computing/the-history-of-pets-vs-cattle/).
- Change the actual hostname of one of the systems so the
    `<username>@<hostname>` pattern works. That is way too out much of
    an ask. `home-manager` shouldn't be influencing the naming of a
    machine.
- Probably some other things.

Regardless, the situation isn't great. We don't do any of those
workarounds. Instead, we do the workaround of moving
`homeConfigurations` (and `homeModules`, since we're already moving
stuff) to `outputs.legacyPackages.<system>.homeConfigurations`:
https://github.com/nix-community/home-manager/issues/3075#issuecomment-1477155995.
This makes it so that the configuration can still work, and because
`home-manager switch` and friends run a `nix build` command (or
something like that), they don't actually need to know that the
`homeConfigurations` is under `outputs.legacyPackages.<system>`, instead
of being under `outputs`.