Install `signal`
Missed adding this one to the `brew-nix` dependencies, primarily because
it's hard to work with `signal`'s ideas of backups. Love the e2e
encryption, and it's gotten better with transferring of backups. But it
still needs some time to smooth out the rough edges.

Regardless, we're adding it now because new computers.