Start managing Firefox more
Turns out, we can manage more of what we do with Firefox in
`home-manager`. We move from pulling it from `brew-nix` to pulling it
from `nixpkgs`, and using the `home-manager` module for it.

This lets us manage things like containers, extensions, policies,
profiles, settings and probably other stuff. We're only doing very
little stuff, but hopefully this can be expanded to more stuff.