Set `NH_DARWIN_FLAKE` and `NH_HOME_FLAKE`
The `nh` module doesn't set these environment variables. It only sets
`NH_FLAKE`. There's a hierachy to this stuff, though. Which is that the
more specific environment variables are checked first. I.e.:
- `NH_DARWIN_FLAKE` is checked first for `nh darwin` commands.
- `NH_HOME_FLAKE` is checked first for `nh home` commands.
- `NH_OS_FLAKE` is checked first for `nh os` commands.

For each of the commands (`nh darwin`, `nh home`, and `nh os`) if the
specific environment variable isn't set, then to falls back to
`NH_FLAKE`.

The issue comes when you have more than one comman you want to run, the
configuration is not in the same flake, and only `NH_FLAKE` is set. E.g.
if you have `NH_FLAKE` set to the `home-manager` location, and run
`nh darwin` commands without having a `darwinConfigurations` in the
flake (like we do in this flake).

The workaround is to set each of the environment variables separately,
like we're doing in this change.

The real solution here seems like it's to have both `homeConfigurations`
and `darwinConfigurations` in the same flake. That seems like it would
be less hassle. Maybe one day…