Make `home.nix` more general
`home.nix` having `home.homeDirectory` and `home.username` explicitly
written in it means that by default the user is setup to be `joneshf`.
This doesn't work once we get to multiple different users.

More to the point, other values using the let-bound `home-directory` and
`username` variables, completely breaks any overriding we'd be able to
do to work around the issue of having `home.homeDirectory` and
`home.username` defined in `home.nix`.

We move these out of the `home.nix` file so they have to be set
elsewhere. We do it in the `flake.nix` file, since that's where the user
is decided anyway. So it makes the most sense.