Add `vscode-fileutils` extension to VSCode
This extension is basically essential for being able to actually work
with files in VSCode. It's kind of appalling that there's not a native
way to create a file at a specific path without using the clunky
Explorer.

This extension isn't in `nixpkgs-unstable`, so we use
`vscode-utils.buildVscodeMarketplaceExtension` to pull it in explicitly.