Use standard conventions for `home.homeDirectory`
`home-manager` needs to know what your home directory is. That makes
sense, because it cannot put files in your home directory if it doesn't
know what that is.

It used to figure out the home directory based on environment variables,
but that behavior was removed in:
https://github.com/nix-community/home-manager/pull/1269. That also makes
sense, relying on environment variables can cause problems, so removing
that part of the behavior seems right.

The replacement was that everyone has to define `home.homeDirectory`
explicitly. That seems a little weird, considering there are standard
conventions for what `home.homeDirectory` should be. `home-manager`
could do this for folks, instead of requring everyone to set the option
themselves. It's not a huge deal to set it explicitly yourself. But it
is another papercut with `home-manager`.

We make this a little easier by adding a small module to set
`home.homeDirectory` based on standard platform conventions. We enable
it by default, because this shouldn't be a thing folks should have to
think about. And we for sure don't want to think about it.

We also move the `home.nix` module from
`outputs.legacyPackages.<system>.homeModules.home` to
`outputs.legacyPackages.<system>.homeModules."home.nix"` so it doesn't
overwrite the file structure and make this
`home.home-directory-convention` module unavailable.