Extract Unfree package behavior to a new module
We want to start splitting up some of this stuff. In order to make that
a little easier, we make a module for Unfree packages. This should let
us have modules that happen to contain Unfree packages in multiple
files instead of needing to define them all in the root `home.nix`
module.