Extract media to a new module
We can move Spotify out of the `home.nix` module now that setting Unfree
packages doesn't have to be done in a single place.

We start a `media` module, since it's kind of its own thing. We put
Spotify in the `media` module (rather than like a "music" module)
because Spotify does more than just music these days. We want to be more
opinionated about Spotify beyond just enabling the program, so it seems
right to have it in this `media` module.

We'll add more to this new module over time.