Extract version control to a new module
Turns out, a non-trivial amount of `home.nix` is about version control.
We pull a bunch of this logic out to its own module so it's a little
clearer how this is supposed to work.

We get pretty opinionated about stuf in the `version-control` module.
Like setting up a bunch of `git` programs, and `jj` stuff. All of this
stuff doesn't have to be used all at once. But having it available to be
used is what's important.