Extract editors to a new module
Similar to the other changes for Unfree packages. We can manage them
better now that we've got Unfree logic in its own package.

We make a new `editors` module and move the JetBrains and VSCode editors
to that module. We're going to be opinionated about these as well, so
having them in their own modules makes sense.