Use `nixpkgs-unstable` as `nixpkgs`
Nearly every package we're using is using `nixpkgs-unstable` right now.
And each time we add a new package, it's to `nixpkgs-unstable`. Even
though `nixpkgs-unstable` is the bleeding edge, we haven't run into any
issues with conflicts and huge failures like we used to with Arch Linux.
So it doesn't seem that dangerous to live on `nixpkgs-unstable`.

We can always change this back, or add in `nixpkgs` at a stable version
for things that break, if it happens. But for now, we make things a
little simpler by not having to explicitly opt-in to the same thing
everywhere.