Use `jetbrains-toolbox` from `nixpkgs`
The version of `jetbrains-toolbox` in `nixpkgs` is a little bit out of
date, but we don't need the absolute latest version anyway.

Bit benefit here is that the package is going to be cached, so we're
only downloading it from Hydra instead of building it ourselves.

Another smaller benefit is that the license is correct for this. That's
why we have to explicitly opt-in with `nixpkgs.unfree-packages`.

In any case, this should be basically the same as before, just with a
slight downgrade in version.