Import GitHub public GPG key
We want to be able to verify signed commits when they're signed by
GitHub. To do that, we need to import the public key of GitHub.

It doesn't seem to be documented what the GPG key is that GitHub uses,
but some amount of spelunking seems to point to
`https://github.com/web-flow.gpg`. It seems to work, so we're going with
it.