Move GPG logic to a new module
There are two GPG modules in `home-manager`, and neither of them does
what we need here. The `programs.gpg` module doesn't set any environment
variables that GPG actually uses in order to verify signatures. The
`services.gpg-agent` doesn't import any public keys so we can verify
signatures; it also doesn't work on macOS (there's a PR that's been open
for ~9 months that sounds like it will fix it:
https://github.com/nix-community/home-manager/pull/5901).

We don't want to wait for this to be settled upstream, so we make our
own module that does what we want it do by gluing together stuff.
Hopefully this story can be made easier in the future, so we don't have
to maintain this module at some point.