Extract finance stuff to a new module
We want to bundle all of this stuff together so it's easier to _not_
install it. We only really need finance stuff on one machine per person
(since we don't generally track finances on servers or whatnot, and
never on work stuff).

Pretty sure `turbotax-2024` is Unfree, but the licenses for `brew-nix`
are questionable.