Extract basic Nix stuff to a module
One of the main problems with Nix is its UX, it's obnoxiously hard to
use. This can be made better by picking up a handful of useful programs
that either simplify what you need to do, make what Nix is doing
clearer, or make things implicit. The problem is knowing what those
useful programs are.

We make a module that sets up a bunch of programs to make Nix actually
usable.

N.B. We don't move `nixpkgs-fmt` into this `usable-nix` module.
`nixpkgs-fmt` was archived about a year ago:
https://github.com/nix-community/nixpkgs-fmt/commit/bdb15b4c7e0cb49ae091dd43113d0a938afae02c.
So we're not going to carry that forward to this new module.