Support macOS "System Settings"
The current state of things with macOS "System Settings" is incredibly
confusing. Nearly everything uses the `defaults` program, and the
domains/keys for `defaults` are almost entirely undocumented. Worse
still, they don't line up with anything you'd see as a human when you
use the "System Settings" UI.

E.g. The setting in "System Settings" for "Accessibility > Zoom > Use
keyboard shortcuts to zoom" is in the domain `com.apple.universalaccess`
with the key `closeViewHotkeysEnabled`. The `close` part of the key is
not the verb "close" (pronounced "klōz") as in the verb you use to exit
a program; but the adjective "close" (pronounced "klōs") as in the
adjective describing distance. It's stuff like this that is incredibly
hard to understand.

Rather than perpetrating this awful UX, we decide to call these options
what they are called in the "System Settings" UI. This is almost surely
going to change when Apple decides to change the copy of these settings
or alter which settings exist. But dealing with that problem is worlds
easier than dealing with the current state of `defaults`.

Regardless, we start by adding support for two settings:
- "Trackpad > Point & Click > Force Click and haptic feedback"
- "Trackpad > Point & Click > Look up & data detectors"

These two settings are coupled, so they kind of go together.