Update flake inputs
It's been quite a while since the last update, so we update everything
we can.

We get a hopefully huge update: `jujutsu` at 0.34.0, which adds support
for string replacements, so we can hopefully finally have human readable
bookmarks generated! But we'll have to see for sure when we try to use
it.

Also, we don't need to pull in `jjui` from upstream anymore, since the
version we wanted is in `nixpkgs-unstable` now. We'll remove that in a
future change.