Make `jujutsu` bookmarks human readable
We can finally have memorable `git` branch names created from `jujutsu`!
This has been a long time coming, but it's so worth it.

I'm sure having the bookmark be some random string of characters is fine
if you never have to interact with branches on like GitHub. But in
practice it makes using `jujutsu` harder than it needs to be because you
have to do a mapping from line noise to something your brain actually
knows about. E.g. having both `joneshf/<line-noise>` and
`joneshf/<slightly-different-line-noise>` doesn't tell you which one is
about feature 1 and which is about feature 2.

With a few bookmarks, it's not too bad. With 5-ish bookmarks in a
megamerge, it's still somewhat manageable. But trying to juggle 10, or
more bookmarks in a megamerge with the bookmark names being line noise
is way too hard. Nevermind trying to share a branch name with
co-workers, they'll definitely never remember which one is which
(because they have even less context since they don't have the change id
themselves). Even more fun when your "co-worker" is actually you on a
different machine.

A few releases ago, we'd always have the line noise of the change id in
the generated bookmarks. An improvement was made to let the bookmark be
generated by a template, so we could control it more:
https://github.com/jj-vcs/jj/pull/6847. Unfortunately, it was next to
impossible to actually generate something useful for the bookmark
because there were no string manipulation functions that worked to break
down a description into something useful as a bookmark.

There was a clever idea to embed some information into the trailers of a
commit, and extract that into the template:
https://github.com/jj-vcs/jj/discussions/6891. This worked okay for
stuff like Jira ids. But it doesn't fix the underlying problem for the
rest of the bookmark name. And unless you only ever work on a single
bookmark per Jira Issue, the template still needs to include some
information to disambiguate two bookmarks for the same Issue. So, we
still had the line noise of the change id in the bookmark.

But now that `jujutsu` 0.34.0 has support for replacing parts of a
string: https://github.com/jj-vcs/jj/pull/7467, we can actually use the
first line of the description and turn it into the generated bookmark
name! So our bookmarks can look like `joneshf/start-work-on-feature-1` and
`joneshf/test-feature-2-more-thoroughly`.

It's kind of wild that this hasn't been a larger
annoyance/blocker/problem for people. There are a lot of folks that get
upset about what seem like very low value issues, but I guess `git`
branch names are even lower value to most people.

In any case, we make the bookmark template able to support a few
different prefixes:
- Jira Issue
- Team name

If none of those match, there's nothing put into the template. This is
all heavily based on how things are done at work, which is a bit of a
shame. But I don't currently have the energy to think about this in a
more general way. Finally, we append the rest of the description's first
line as the suffix to the bookmark.

The implementation is actually pretty complex, but that's more to do
with templates still not having great functions for string manipulation.
It's not the end of the world, and it'll probably get better in the
future if more stuff is added, but this works for now!

As for what it looks like in practice. Given some assortment of changes:
```
$ jj log --no-graph --revisions='oum | os | vp | w' --template='"change_id: " ++ pad_end(3, change_id.shortest()) ++ " description: " ++ json(description.first_line()) ++ "\n"'
change_id: vp  description: "  \"delete\" this-later-!     "
change_id: oum description: "REBELS Harmonic Convergence"
change_id: w   description: "Do this and/or that"
change_id: os  description: "VGI-171 Do the thing!"
```

We can have the bookmarks be generated to human readable names:
```
$ jj git push --change='oum | os | vp | w' --dry-run
Creating bookmark joneshf/delete-this-later for revision vpwotynooquo
Creating bookmark joneshf/REBELS/harmonic-convergence for revision oumtkuwomwkl
Creating bookmark joneshf/do-this-and-or-that for revision wuxtsklqvupn
Creating bookmark joneshf/VGI-171/do-the-thing for revision osvtwpxxtozt
Changes to push to origin:
  Add bookmark joneshf/delete-this-later to 756374db6ea6
  Add bookmark joneshf/REBELS/harmonic-convergence to 3f282ee0ed86
  Add bookmark joneshf/do-this-and-or-that to 44e404ecc14a
  Add bookmark joneshf/VGI-171/do-the-thing to ed484cf3a1e9
Dry-run requested, not pushing.
```