Support private commits with `jj`
There are sometimes changes that aren't ready to be made public. We want
to start a convention that all private stuff is prefixed with
`[PRIVATE]`. We do that, but have to make sure to escape things properly
so it doesn't match the regex of `[PRIVATE]` (which is one of the values
`P`, or `R`, or `I`, or `V`, or `A`, or `T`, or `E`).