Update everything
It's been a minute since everything needed to be updated at once.
Unfortunately the sweater started coming apart, so we had to do it in
this change.

There are only two real regressions here: `krita` and
`python-vipaccess`. Both are explained in a comment around the package.
Hopefully, we can sort this out a some point in the future. But for now,
it's not a big enough deal to spend more time on addressing it than we
already have.

The changes to `programs.ssh` are to silence the warning about the
deprecation of default values. This should be a no-op. But we're calling
it out anyway.

In any case, we get some updates in this change including `jj` and
`jjui`. The latter finally has support for megamerges!

Also, the 1password shell plugin flake has finally updated to remove the
warning from `nixpkgs`!